Added test controller and db_view to see everything in the databas quickly without queries

+++ app/controllers/test_controller.rb
@@ -3,7 +3,7 @@ class TestController < ApplicationController
   def db_view
     table_html = ''
     table_names = ActiveRecord::Base.connection.tables.delete_if { |x| x == 'schema_migrations' }
-    table_names.each do |name|
+    table_names.sort.each do |name|
       table_name = "#{name.titleize.gsub(' ', '').singularize}"
       table_data = eval("#{table_name}.all")
 
@@ -13,7 +13,7 @@ class TestController < ApplicationController
         row.attributes.each do |column_name, column_data|
           table_html << "<li>#{column_name} : #{column_data}</li>"
         end
-        table_html << '</ul>'
+        table_html << "<li><a href='/test/delete_db_view_record/#{table_name}_#{row.id}' class='btn btn-large btn-primary' style='font-size:20px;'>[Delete Record]</a></li></ul>"
       end
 
     end
@@ -24,6 +24,17 @@ class TestController < ApplicationController
     if not params[:id].blank?
       data = eval("#{params[:id]}.all")
       data.each do |record|
+        record.destroy
+      end
+    end
+    redirect_to :action => 'db_view'
+  end
+
+  def delete_db_view_record
+    if not params[:id].blank?
+      table_name, row_id = params[:id].split('_')
+      record = eval("#{table_name}.first(:conditions => ['id = ?', #{row_id}])")
+      if record
         record.destroy
       end
     end
